refactor User action/reducer to FetchUser
