+++ src/Reducers/RootReducer.js
@@ -1,5 +1,5 @@
 import {combineReducers} from 'redux';
-import user from './UserReducer';
+import user from './FetchUserReducer';
 
 export default combineReducers({
     user
